Rename solr.update_work to solr.update
