+++ openlibrary/plugins/openlibrary/dev_instance.py
@@ -114,7 +114,7 @@ class process_ebooks(delegate.page):
 @oltask
 def update_solr(changeset):
     """Updates solr on edit."""
-    from openlibrary.solr import update_work
+    from openlibrary.solr import update
 
     keys = set()
     docs = changeset['docs'] + changeset['old_docs']
@@ -130,7 +130,7 @@ def update_solr(changeset):
         elif doc['type']['key'] == '/type/author':
             keys.add(doc['key'])
 
-    update_work.update_keys(list(keys))
+    update.update_keys(list(keys))
 
 
 @infogami.install_hook
